refactor: async CRUD

+++ componentSql/user.py
@@ -4,7 +4,10 @@ from typing import List
 
 import sqlalchemy
 from sqlalchemy import Column, Integer, String, text
+from sqlalchemy.exc import IntegrityError
+
 from core.sqlEngine import sql_engine
+from sqlalchemy.future import select
 
 
 class UserInfo(sql_engine.SqlBase):
@@ -34,34 +37,42 @@ class UserInfo(sql_engine.SqlBase):
         return UserInfo.select_by_username(userinfo.username)
 
     @staticmethod
-    def select_by_username(username):
-        with sql_engine.get_session() as session:
-            res = session.query(UserInfo).filter(UserInfo.username == username).first()
-            return res
+    async def select_by_username(username):
+        print(username)
+        async with sql_engine.get_async_session() as session:
+            # 使用异步查询
+            query = select(UserInfo).filter(UserInfo.username == username)
+            result = await session.execute(query)
+            user = result.scalars().first()
+            return user
 
     @staticmethod
-    def create_user(userinfo: UserInfo):
-        with sql_engine.get_session() as session:
+    async def create_user(username, password):
+        userinfo = UserInfo(username=username, password=password)
+        async with sql_engine.get_async_session() as session:
             try:
                 session.add(userinfo)
-                session.commit()
-                # 可选：刷新会话以获取数据库生成的值（如自动递增的ID）
-                session.refresh(userinfo)
+                await session.commit()  # 异步提交
+                await session.refresh(userinfo)  # 异步刷新
                 sql_engine.sql_log.info(f"Created user: {userinfo}")
-                return userinfo  # 可选返回创建的用户对象
-            except sqlalchemy.exc.IntegrityError as e:
+                return userinfo  # 返回创建的用户对象
+            except IntegrityError as e:
                 sql_engine.sql_log.exception(e)
                 return False
 
     @staticmethod
-    def del_user(userinfo: UserInfo):
-        userinfo = userinfo.select_by_username(userinfo.username)
-        with sql_engine.get_session() as session:
-            try:
+    async def del_user(username):
+        # 异步查询用户对象
+        userinfo = await UserInfo.select_by_username(username)
+        if userinfo is None:
+            sql_engine.sql_log.info(f"User not found: {username}")
+            return False  # 用户不存在，返回 False
 
+        async with sql_engine.get_async_session() as session:
+            try:
                 # 删除指定的用户对象
-                session.delete(userinfo)
-                session.commit()
+                await session.delete(userinfo)
+                await session.commit()
                 sql_engine.sql_log.info(f"Deleted user: {userinfo}")
                 return True  # 返回 True 表示删除成功
             except Exception as e:
@@ -78,25 +89,8 @@ class UserInfo(sql_engine.SqlBase):
 
 
 if __name__ == '__main__':
-    from sqlalchemy import select
+    # userinfo = UserInfo(username='123', password='123')
+    # userinfo.create_user(username='456', password='123')
     import asyncio
-
-
-    async def get_users():
-        async with sql_engine.get_async_session() as session:
-            async with session.begin():
-                # 构建查询语句
-                query = select(UserInfo).filter(UserInfo.username == 'mqnu00')
-                # 执行查询
-                result = await session.execute(query)
-                # 解析结果
-                users = result.scalars().all()
-                for user in users:
-                    print(user)
-
-
     loop = asyncio.get_event_loop()
-    loop.run_until_complete(get_users())
-
-    print(UserInfo.select_by_username('mqnu00'))
-
+    loop.run_until_complete(UserInfo().del_user('123'))
